+++ src/util/factor.cpp
@@ -36,41 +36,41 @@
 
 using namespace schnek;
 
-void makePrimes(std::list<int> &primes, int max);
+void makePrimes(std::list<size_t> &primes, size_t max);
 
-void factorize(int number, std::list<int> &primes, std::list<int> &factors);
+void factorize(size_t number, std::list<size_t> &primes, std::list<size_t> &factors);
 
-double distribute(std::vector<int> &factors, std::list<int> allfact, std::vector<int> &weights);
+double distribute(std::vector<size_t> &factors, std::list<size_t> allfact, std::vector<size_t> &weights);
 
-void schnek::equalFactors(int number, int nfact, std::vector<int> &factors, std::vector<int> &weights) {
-  std::list<int> primes;
-  std::list<int> allfact;
-  makePrimes(primes, int(floor(sqrt(number))));
+void schnek::equalFactors(size_t number, size_t nfact, std::vector<size_t> &factors, std::vector<size_t> &weights) {
+  std::list<size_t> primes;
+  std::list<size_t> allfact;
+  makePrimes(primes, size_t(floor(sqrt(number))));
   factorize(number, primes, allfact);
 
   factors.resize(nfact);
 
   // this happens only on single processor runs
   if (allfact.empty()) {
-    for (int i = 0; i < nfact; ++i) factors[i] = 1;
+    for (size_t i = 0; i < nfact; ++i) factors[i] = 1;
     return;
   }
 
   factors[0] = allfact.front();
-  for (int i = 1; i < nfact; ++i) factors[i] = 1;
+  for (size_t i = 1; i < nfact; ++i) factors[i] = 1;
 
   allfact.pop_front();
   if (allfact.empty()) return;
   distribute(factors, allfact, weights);
 }
 
-void makePrimes(std::list<int> &primes, int max) {
+void makePrimes(std::list<size_t> &primes, size_t max) {
   std::vector<bool> isprime(max + 1, true);
   primes.clear();
-  for (int i = 2; i <= max; ++i) {
+  for (size_t i = 2; i <= max; ++i) {
     if (isprime[i]) {
       primes.push_back(i);
-      int prod = 2 * i;
+      size_t prod = 2 * i;
       for (prod = 2 * i; prod <= max; prod += i) {
         isprime[prod] = false;
       }
@@ -78,12 +78,12 @@ void makePrimes(std::list<int> &primes, int max) {
   }
 }
 
-void factorize(int number, std::list<int> &primes, std::list<int> &factors) {
+void factorize(size_t number, std::list<size_t> &primes, std::list<size_t> &factors) {
   if (number == 1) return;
-  std::list<int>::iterator it = primes.begin();
+  std::list<size_t>::iterator it = primes.begin();
 
   while (it != primes.end()) {
-    int p = *it;
+    size_t p = *it;
     if ((number % p) == 0) {
       factorize(number / p, primes, factors);
       factors.push_front(p);
@@ -94,15 +94,15 @@ void factorize(int number, std::list<int> &primes, std::list<int> &factors) {
   factors.push_front(number);
 }
 
-double distribute(std::vector<int> &factors, std::list<int> allfact, std::vector<int> &weights) {
-  int f = allfact.front();
+double distribute(std::vector<size_t> &factors, std::list<size_t> allfact, std::vector<size_t> &weights) {
+  size_t f = allfact.front();
   allfact.pop_front();
 
   if (allfact.empty()) {
     std::vector<double> dfactors;
     dfactors.resize(factors.size());
 
-    int imin = 0;
+    size_t imin = 0;
     double facmin = 0;
 
     for (size_t i = 0; i < factors.size(); ++i) {
@@ -119,12 +119,12 @@ double distribute(std::vector<int> &factors, std::list<int> allfact, std::vector
     return std::for_each(dfactors.begin(), dfactors.end(), calcsum<double>()).result();
 
   } else {
-    std::vector<int> best_factors = factors;
+    std::vector<size_t> best_factors = factors;
     best_factors[0] *= f;
     double bestSum = distribute(best_factors, allfact, weights);
 
     for (size_t i = 1; i < factors.size(); ++i) {
-      std::vector<int> t_factors = factors;
+      std::vector<size_t> t_factors = factors;
       t_factors[i] *= f;
       double sum = distribute(t_factors, allfact, weights);
       if (sum < bestSum) {
@@ -142,11 +142,11 @@ double distribute(std::vector<int> &factors, std::list<int> allfact, std::vector
 // Testing program for factors
 int main()
 {
-  std::vector<int> factors;
-  int number = 64;
-  int nfact = 3;
+  std::vector<size_t> factors;
+  size_t number = 64;
+  size_t nfact = 3;
 
-  std::vector<int> weights(nfact);
+  std::vector<size_t> weights(nfact);
   weights[0] = 32;
   weights[1] = 32;
   weights[2] = 256;
@@ -154,7 +154,7 @@ int main()
   equalFactors(number, 3, factors, weights);
 
   std::cout << number << " = ";
-  for (int i = 0; i < nfact; ++i)
+  for (size_t i = 0; i < nfact; ++i)
   {
     std::cout << factors[i] << "*";
   }
